+++ plugin.php
@@ -13,7 +13,7 @@
  * Plugin URI:  https://wordpress.org/plugins/wp-plugin-dependencies
  * Description: Parses 'Requires Plugins' header, add plugin install dependencies tab, and information about dependencies.
  * Author: Andy Fragen, Colin Stewart
- * Version: 1.4.1
+ * Version: 1.5.0
  * License: MIT
  * Network: true
  * Requires at least: 6.0
@@ -37,6 +37,19 @@ if ( version_compare( get_bloginfo( 'version' ), '6.1-RC1', '>=' ) ) {
 	deactivate_plugins( __FILE__ );
 }
 
+// Load the Composer autoloader.
+if ( file_exists( __DIR__ . '/vendor/autoload.php' ) ) {
+	require __DIR__ . '/vendor/autoload.php';
+} else {
+	\deactivate_plugins( __FILE__ );
+
+	wp_die(
+		wp_kses_post(
+			__( 'Plugin Dependencies is missing required composer dependencies.', 'wp-plugin-dependencies' )
+		)
+	);
+}
+
 /**
  * Class Init
  */
@@ -52,6 +65,17 @@ class Init {
 
 		add_filter( 'install_plugins_tabs', array( $this, 'add_install_tab' ), 10, 1 );
 		add_filter( 'install_plugins_table_api_args_dependencies', array( $this, 'add_install_dependency_args' ), 10, 1 );
+
+		// Add the sites with REST enpoints that return plugins_api() data when passed `slug` query arg.
+		add_filter(
+			'plugin_dependency_endpoints',
+			function () {
+				return array(
+					'https://git-updater.com/wp-json/git-updater/v1/plugins-api/',
+				);
+			}
+		);
+
 		add_action( 'install_plugins_dependencies', 'display_plugins_table' );
 		add_action( 'admin_init', array( 'WP_Plugin_Dependencies', 'init' ) );
 	}