+++ test-plugins/test-dependencies2.php
@@ -8,5 +8,5 @@
  * Author: WordPress Core Contributors
  * Version: 0.1
  * Description: Testing inclusion of non-dot org plugins in the Dependencies tab.
- * Requires Plugins: gravity-forms
+ * Requires Plugins: gravity-forms, git-updater
  */