+++ test-plugins/test-dependencies3.php
@@ -8,5 +8,5 @@
  * Author: WordPress Core Contributors
  * Version: 0.1
  * Description: Testing inclusion of dot org plugins in the Dependencies tab.
- * Requires Plugins: gutenberg, wp-plugin-dependencies, hello-dolly
+ * Requires Plugins: gutenberg, wp-plugin-dependencies, akismet
  */